Avoid unnecessary rendering of App component
