+++ mini-redux/src/useAppState.ts
@@ -1,4 +1,5 @@
-import { createContext, useContext } from "react";
+import { createContext } from "react";
+import { useAppStateStore } from "./useAppStateStore";
 
 export interface AppState {
   articles: {
@@ -11,10 +12,8 @@ export interface AppState {
 export const appStateContext = createContext<AppState | null>(null);
 
 export function useAppState() {
-  const appState = useContext(appStateContext);
-  if (!appState) {
-    throw new Error("Provider で囲んでください");
-  }
+  const store = useAppStateStore();
+  const appState = store.getState();
 
   return appState;
 }
